+++ src/gen/io.linkerd.proxy.identity.rs
@@ -1,0 +1,244 @@
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct CertifyRequest {
+    #[prost(string, tag="1")]
+    pub identity: ::prost::alloc::string::String,
+    /// Proof of the requester's identity.
+    ///
+    /// In Kubernetes, for instance, this is the contents of a service account
+    /// token.
+    #[prost(bytes="vec", tag="2")]
+    pub token: ::prost::alloc::vec::Vec<u8>,
+    /// A PEM-encoded x509 Certificate Signing Request.
+    #[prost(bytes="vec", tag="3")]
+    pub certificate_signing_request: ::prost::alloc::vec::Vec<u8>,
+}
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct CertifyResponse {
+    /// A PEM-encoded x509 Certificate.
+    #[prost(bytes="vec", tag="1")]
+    pub leaf_certificate: ::prost::alloc::vec::Vec<u8>,
+    /// A list of PEM-encoded x509 Certificates that establish the trust chain
+    /// between the leaf_certificate and the well-known trust anchors.
+    #[prost(bytes="vec", repeated, tag="2")]
+    pub intermediate_certificates: ::prost::alloc::vec::Vec<::prost::alloc::vec::Vec<u8>>,
+    #[prost(message, optional, tag="3")]
+    pub valid_until: ::core::option::Option<::prost_types::Timestamp>,
+}
+/// Generated client implementations.
+pub mod identity_client {
+    #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
+    use tonic::codegen::*;
+    #[derive(Debug, Clone)]
+    pub struct IdentityClient<T> {
+        inner: tonic::client::Grpc<T>,
+    }
+    impl<T> IdentityClient<T>
+    where
+        T: tonic::client::GrpcService<tonic::body::BoxBody>,
+        T::Error: Into<StdError>,
+        T::ResponseBody: Default + Body<Data = Bytes> + Send + 'static,
+        <T::ResponseBody as Body>::Error: Into<StdError> + Send,
+    {
+        pub fn new(inner: T) -> Self {
+            let inner = tonic::client::Grpc::new(inner);
+            Self { inner }
+        }
+        pub fn with_interceptor<F>(
+            inner: T,
+            interceptor: F,
+        ) -> IdentityClient<InterceptedService<T, F>>
+        where
+            F: tonic::service::Interceptor,
+            T: tonic::codegen::Service<
+                http::Request<tonic::body::BoxBody>,
+                Response = http::Response<
+                    <T as tonic::client::GrpcService<tonic::body::BoxBody>>::ResponseBody,
+                >,
+            >,
+            <T as tonic::codegen::Service<
+                http::Request<tonic::body::BoxBody>,
+            >>::Error: Into<StdError> + Send + Sync,
+        {
+            IdentityClient::new(InterceptedService::new(inner, interceptor))
+        }
+        /// Compress requests with `gzip`.
+        ///
+        /// This requires the server to support it otherwise it might respond with an
+        /// error.
+        #[must_use]
+        pub fn send_gzip(mut self) -> Self {
+            self.inner = self.inner.send_gzip();
+            self
+        }
+        /// Enable decompressing responses with `gzip`.
+        #[must_use]
+        pub fn accept_gzip(mut self) -> Self {
+            self.inner = self.inner.accept_gzip();
+            self
+        }
+        /// Requests that a time-bounded certificate be signed.
+        ///
+        /// The requester must provide a token that verifies the client's identity and
+        /// a Certificate Signing Request that adheres to the service naming rules.
+        ///
+        /// Errors are returned when the provided request is invalid or when
+        /// authentication cannot be performed.
+        pub async fn certify(
+            &mut self,
+            request: impl tonic::IntoRequest<super::CertifyRequest>,
+        ) -> Result<tonic::Response<super::CertifyResponse>, tonic::Status> {
+            self.inner
+                .ready()
+                .await
+                .map_err(|e| {
+                    tonic::Status::new(
+                        tonic::Code::Unknown,
+                        format!("Service was not ready: {}", e.into()),
+                    )
+                })?;
+            let codec = tonic::codec::ProstCodec::default();
+            let path = http::uri::PathAndQuery::from_static(
+                "/io.linkerd.proxy.identity.Identity/Certify",
+            );
+            self.inner.unary(request.into_request(), path, codec).await
+        }
+    }
+}
+/// Generated server implementations.
+pub mod identity_server {
+    #![allow(unused_variables, dead_code, missing_docs, clippy::let_unit_value)]
+    use tonic::codegen::*;
+    ///Generated trait containing gRPC methods that should be implemented for use with IdentityServer.
+    #[async_trait]
+    pub trait Identity: Send + Sync + 'static {
+        /// Requests that a time-bounded certificate be signed.
+        ///
+        /// The requester must provide a token that verifies the client's identity and
+        /// a Certificate Signing Request that adheres to the service naming rules.
+        ///
+        /// Errors are returned when the provided request is invalid or when
+        /// authentication cannot be performed.
+        async fn certify(
+            &self,
+            request: tonic::Request<super::CertifyRequest>,
+        ) -> Result<tonic::Response<super::CertifyResponse>, tonic::Status>;
+    }
+    #[derive(Debug)]
+    pub struct IdentityServer<T: Identity> {
+        inner: _Inner<T>,
+        accept_compression_encodings: (),
+        send_compression_encodings: (),
+    }
+    struct _Inner<T>(Arc<T>);
+    impl<T: Identity> IdentityServer<T> {
+        pub fn new(inner: T) -> Self {
+            Self::from_arc(Arc::new(inner))
+        }
+        pub fn from_arc(inner: Arc<T>) -> Self {
+            let inner = _Inner(inner);
+            Self {
+                inner,
+                accept_compression_encodings: Default::default(),
+                send_compression_encodings: Default::default(),
+            }
+        }
+        pub fn with_interceptor<F>(
+            inner: T,
+            interceptor: F,
+        ) -> InterceptedService<Self, F>
+        where
+            F: tonic::service::Interceptor,
+        {
+            InterceptedService::new(Self::new(inner), interceptor)
+        }
+    }
+    impl<T, B> tonic::codegen::Service<http::Request<B>> for IdentityServer<T>
+    where
+        T: Identity,
+        B: Body + Send + 'static,
+        B::Error: Into<StdError> + Send + 'static,
+    {
+        type Response = http::Response<tonic::body::BoxBody>;
+        type Error = std::convert::Infallible;
+        type Future = BoxFuture<Self::Response, Self::Error>;
+        fn poll_ready(
+            &mut self,
+            _cx: &mut Context<'_>,
+        ) -> Poll<Result<(), Self::Error>> {
+            Poll::Ready(Ok(()))
+        }
+        fn call(&mut self, req: http::Request<B>) -> Self::Future {
+            let inner = self.inner.clone();
+            match req.uri().path() {
+                "/io.linkerd.proxy.identity.Identity/Certify" => {
+                    #[allow(non_camel_case_types)]
+                    struct CertifySvc<T: Identity>(pub Arc<T>);
+                    impl<T: Identity> tonic::server::UnaryService<super::CertifyRequest>
+                    for CertifySvc<T> {
+                        type Response = super::CertifyResponse;
+                        type Future = BoxFuture<
+                            tonic::Response<Self::Response>,
+                            tonic::Status,
+                        >;
+                        fn call(
+                            &mut self,
+                            request: tonic::Request<super::CertifyRequest>,
+                        ) -> Self::Future {
+                            let inner = self.0.clone();
+                            let fut = async move { (*inner).certify(request).await };
+                            Box::pin(fut)
+                        }
+                    }
+                    let accept_compression_encodings = self.accept_compression_encodings;
+                    let send_compression_encodings = self.send_compression_encodings;
+                    let inner = self.inner.clone();
+                    let fut = async move {
+                        let inner = inner.0;
+                        let method = CertifySvc(inner);
+                        let codec = tonic::codec::ProstCodec::default();
+                        let mut grpc = tonic::server::Grpc::new(codec)
+                            .apply_compression_config(
+                                accept_compression_encodings,
+                                send_compression_encodings,
+                            );
+                        let res = grpc.unary(method, req).await;
+                        Ok(res)
+                    };
+                    Box::pin(fut)
+                }
+                _ => {
+                    Box::pin(async move {
+                        Ok(
+                            http::Response::builder()
+                                .status(200)
+                                .header("grpc-status", "12")
+                                .header("content-type", "application/grpc")
+                                .body(empty_body())
+                                .unwrap(),
+                        )
+                    })
+                }
+            }
+        }
+    }
+    impl<T: Identity> Clone for IdentityServer<T> {
+        fn clone(&self) -> Self {
+            let inner = self.inner.clone();
+            Self {
+                inner,
+                accept_compression_encodings: self.accept_compression_encodings,
+                send_compression_encodings: self.send_compression_encodings,
+            }
+        }
+    }
+    impl<T: Identity> Clone for _Inner<T> {
+        fn clone(&self) -> Self {
+            Self(self.0.clone())
+        }
+    }
+    impl<T: std::fmt::Debug> std::fmt::Debug for _Inner<T> {
+        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+            write!(f, "{:?}", self.0)
+        }
+    }
+}